ref: Move MangaExtensions to util.manga
Moves `MangaExtensions.kt` from `eu.kanade.tachiyomi.util` to `eu.kanade.tachiyomi.util.manga` to improve package cohesion and organization. Updates all relevant imports.
