+++ app/src/main/java/eu/kanade/tachiyomi/ui/feed/FeedRepository.kt
@@ -13,8 +13,8 @@ import eu.kanade.tachiyomi.data.download.model.Download
 import eu.kanade.tachiyomi.ui.manga.MangaConstants
 import eu.kanade.tachiyomi.util.chapter.ChapterItemSort
 import eu.kanade.tachiyomi.util.chapter.isAvailable
+import eu.kanade.tachiyomi.util.manga.toDisplayManga
 import eu.kanade.tachiyomi.util.system.executeOnIO
-import eu.kanade.tachiyomi.util.toDisplayManga
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Date
