+++ app/src/main/java/eu/kanade/tachiyomi/util/manga/MangaShortcutManager.kt
@@ -16,7 +16,6 @@ import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.ui.feed.FeedRepository
 import eu.kanade.tachiyomi.ui.main.MainActivity
 import eu.kanade.tachiyomi.util.system.launchIO
-import eu.kanade.tachiyomi.util.toDisplayManga
 import kotlin.math.min
 import kotlinx.coroutines.GlobalScope
 import org.nekomanga.R
